Changing the creation of swift_defconfig to use a patch file
This commit updates how the swift_defconfig file is created.  Now
the swift_setup script will copy down the latest witherspoon_defconfig
file.  It will rename the file to swift_defconfig and then apply the patch
file. This allows for any overall (ie. not-swift-specific) changes
made to witherspoon_defconfig at the op-build level to be picked up
automatically for the swift_defconfig.

Signed-off-by: Mike Baiocchi <mbaiocch@us.ibm.com>