Allow setting up the swift config files in an existing op-build
This commit updates the ./setup_swift script to allow for a user
to pass in an existing op-build directory that they want the script
to setup the swift config files in.  In this case, cloning a new
op-build is not done.  It also allows for a user to skip linking
the config files to the op-build directory if they want.